requirements.txt
protobuf<=3.20.3
absl-py==2.1.0
boto3==1.33.13
Flask==2.2.5
numpy==1.18.5
six==1.17.0
tensorflow==1.15.5
gunicorn